pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-openai>=0.3.23",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "supabase>=2.15.3",
    "uvicorn[standard]>=0.34.3",
]
